updating previously cached files when they are updated in the backend.
when re-mounting mcachefs with previously cached files, originally mcachefs_transfer_do_backing detects a file already in cache and would throw an error, forcing mcachefs to always use the source file instead of updating the local cache. 

This change just reports that the already cached file exists and it's older than the source, and instead of returning and forcing mcachefs to use the source file, it will continue and update the file in the cache, so subsequent reading will be done from the cache as it should.